Refactor oesign signing functions out of main.c
- Move all signing related functions out of main.c into new oesign.c file.
- Clean up header includes in refactored main.c, oesign.c and oedump.c.
